Add filters to logging (#966)
* Add filters to logging, update docs

* Remove commented out code from __main__

* Refactor logging, add tests

* Remove unused imports - linting fix

* Remove commented out code

* Handle whitelist and blacklist filter used together. Update docs

* Remove test print

* Refactor logging filter code
