+++ opsdroid/__main__.py
@@ -6,13 +6,13 @@ import sys
 import logging
 import gettext
 import time
-import contextlib
 
 import click
 
 from opsdroid import __version__
 from opsdroid.core import OpsDroid
 from opsdroid.loader import Loader
+from opsdroid.logging import configure_logging
 from opsdroid.const import (
     DEFAULT_LOG_FILENAME,
     LOCALE_DIR,
@@ -39,81 +39,6 @@ def configure_lang(config):
     if lang_code != DEFAULT_LANGUAGE:
         lang = gettext.translation("opsdroid", LOCALE_DIR, (lang_code,), fallback=True)
         lang.install()
-
-
-def configure_logging(config):
-    """Configure the root logger based on user config."""
-    rootlogger = logging.getLogger()
-    while rootlogger.handlers:
-        rootlogger.handlers.pop()
-
-    try:
-        if config["logging"]["path"]:
-            logfile_path = os.path.expanduser(config["logging"]["path"])
-        else:
-            logfile_path = config["logging"]["path"]
-    except KeyError:
-        logfile_path = DEFAULT_LOG_FILENAME
-
-    try:
-        log_level = get_logging_level(config["logging"]["level"])
-    except KeyError:
-        log_level = logging.INFO
-
-    rootlogger.setLevel(log_level)
-    formatter = logging.Formatter("%(levelname)s %(name)s: %(message)s")
-
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(log_level)
-    console_handler.setFormatter(formatter)
-    rootlogger.addHandler(console_handler)
-
-    with contextlib.suppress(KeyError):
-        if not config["logging"]["console"]:
-            console_handler.setLevel(logging.CRITICAL)
-
-    if logfile_path:
-        logdir = os.path.dirname(os.path.realpath(logfile_path))
-        if not os.path.isdir(logdir):
-            os.makedirs(logdir)
-        file_handler = logging.FileHandler(logfile_path)
-        file_handler.setLevel(log_level)
-        file_handler.setFormatter(formatter)
-        rootlogger.addHandler(file_handler)
-    _LOGGER.info("=" * 40)
-    _LOGGER.info(_("Started opsdroid %s"), __version__)
-
-
-def get_logging_level(logging_level):
-    """Get the logger level based on the user configuration.
-
-    Args:
-        logging_level: logging level from config file
-
-    Returns:
-        logging LEVEL ->
-            CRITICAL = 50
-            FATAL = CRITICAL
-            ERROR = 40
-            WARNING = 30
-            WARN = WARNING
-            INFO = 20
-            DEBUG = 10
-            NOTSET = 0
-
-    """
-    if logging_level == "critical":
-        return logging.CRITICAL
-
-    if logging_level == "error":
-        return logging.ERROR
-    if logging_level == "warning":
-        return logging.WARNING
-
-    if logging_level == "debug":
-        return logging.DEBUG
-
-    return logging.INFO
 
 
 def check_dependencies():
